+++ src/util/index.ts
@@ -25,6 +25,7 @@ export * from "./DistributionListUtils.js";
 export * from "./DraftBodyRetentionUtils.js";
 export * from "./DnsSetupUtils.js";
 export * from "./DomainUtils.js";
+export * from "./TimeZoneUtils.js";
 export * from "./DomainVerificationUtils.js";
 export * from "./EntityUtils.js";
 export * from "./FocusedInboxUtils.js";